+++ src/main/java/com/jocoos/mybeautip/store/StoreRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 
-public interface StoreRepository extends JpaRepository<Store, Long> {
+public interface StoreRepository extends JpaRepository<Store, Integer> {
   @Modifying
   @Query("update Store s set s.likeCount = s.likeCount + ?2, s.modifiedAt = now() where s.id = ?1")
-  void updateLikeCount(Long id, Integer count);
+  void updateLikeCount(Integer id, Integer count);
 }
